// src/transaction/types.ts
export interface SendTransactionResult {
  /** The transaction */
  transaction: Transaction
  /** The response if the transaction was sent and waited for */
  confirmation?: PendingTransactionResponse
}
// ...
export type SendTransactionFrom = AddressWithTransactionSigner